refactor: component index
