+++ packages/oruga-next/src/components/button/index.ts
@@ -4,10 +4,12 @@ import Button from "./Button.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export button plugin */
 export default {
     install(app: App) {
         registerComponent(app, Button);
     },
 } as Plugin;
 
+/** export button components */
 export { Button as OButton };
