+++ packages/oruga-next/src/components/checkbox/index.ts
@@ -4,10 +4,12 @@ import Checkbox from "./Checkbox.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export checkbox plugin */
 export default {
     install(app: App) {
         registerComponent(app, Checkbox);
     },
 } as Plugin;
 
+/** export table components */
 export { Checkbox as OCheckbox };
