+++ packages/oruga-next/src/components/datepicker/index.ts
@@ -6,10 +6,12 @@ export type { DatepickerEvent, FocusedDate } from "./useDatepickerShare";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export datepicker plugin */
 export default {
     install(app: App) {
         registerComponent(app, Datepicker);
     },
 } as Plugin;
 
+/** export datepicker components */
 export { Datepicker as ODatepicker };
