+++ packages/oruga-next/src/components/datetimepicker/index.ts
@@ -4,10 +4,12 @@ import Datetimepicker from "./Datetimepicker.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export datetimepicker plugin */
 export default {
     install(app: App) {
         registerComponent(app, Datetimepicker);
     },
 } as Plugin;
 
+/** export datetimepicker components */
 export { Datetimepicker as ODatetimepicker };
