+++ packages/oruga-next/src/components/dropdown/index.ts
@@ -5,6 +5,7 @@ import DropdownItem from "./DropdownItem.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export dropdown plugin */
 export default {
     install(app: App) {
         registerComponent(app, Dropdown);
@@ -12,4 +13,5 @@ export default {
     },
 } as Plugin;
 
+/** export dropdown components */
 export { Dropdown as ODropdown, DropdownItem as ODropdownItem };
