+++ packages/oruga-next/src/components/icon/index.ts
@@ -4,10 +4,12 @@ import Icon from "./Icon.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export icon plugin */
 export default {
     install(app: App) {
         registerComponent(app, Icon);
     },
 } as Plugin;
 
+/** export icon components */
 export { Icon as OIcon };
