+++ packages/oruga-next/src/components/loading/types.ts
@@ -1,0 +1,3 @@
+import Loading from "./Loading.vue";
+
+export type LoadingProps = InstanceType<typeof Loading>["$props"];
