+++ packages/oruga-next/src/components/modal/types.ts
@@ -1,0 +1,3 @@
+import type Modal from "./Modal.vue";
+
+export type ModalProps = InstanceType<typeof Modal>["$props"];
