+++ packages/oruga-next/src/components/notification/types.ts
@@ -1,0 +1,8 @@
+import Notification from "./Notification.vue";
+import NotificationNotice from "./NotificationNotice.vue";
+
+export type NotifcationProps = InstanceType<typeof Notification>["$props"];
+
+export type NotifcationNoticeProps = InstanceType<
+    typeof NotificationNotice
+>["$props"];
