+++ packages/oruga-next/src/components/radio/index.ts
@@ -4,10 +4,12 @@ import Radio from "./Radio.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export radio plugin */
 export default {
     install(app: App) {
         registerComponent(app, Radio);
     },
 } as Plugin;
 
+/** export radio components */
 export { Radio as ORadio };
