+++ packages/oruga-next/src/components/sidebar/types.ts
@@ -1,0 +1,3 @@
+import Sidebar from "./Sidebar.vue";
+
+export type SidebarProps = InstanceType<typeof Sidebar>["$props"];
