+++ packages/oruga-next/src/components/skeleton/index.ts
@@ -4,10 +4,12 @@ import Skeleton from "./Skeleton.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export skeleton plugin */
 export default {
     install(app: App) {
         registerComponent(app, Skeleton);
     },
 } as Plugin;
 
+/** export skeleton components */
 export { Skeleton as OSkeleton };
