+++ packages/oruga-next/src/components/slider/index.ts
@@ -5,7 +5,7 @@ import SliderTick from "./SliderTick.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-// export slider plugin
+/** export slider plugin */
 export default {
     install(app: App) {
         registerComponent(app, Slider);
@@ -13,5 +13,5 @@ export default {
     },
 } as Plugin;
 
-// export slider components
+/** export slider components */
 export { Slider as OSlider, SliderTick as OSliderTick };
