+++ packages/oruga-next/src/components/tooltip/index.ts
@@ -2,12 +2,14 @@ import type { App, Plugin } from "vue";
 
 import Tooltip from "./Tooltip.vue";
 
-import { registerComponent } from "../../utils/plugins";
+import { registerComponent } from "@/utils/plugins";
 
+/** export tooltip plugin */
 export default {
     install(app: App) {
         registerComponent(app, Tooltip);
     },
 } as Plugin;
 
+/** export tooltip components */
 export { Tooltip as OTooltip };
