+++ packages/oruga-next/src/components/upload/index.ts
@@ -4,10 +4,12 @@ import Upload from "./Upload.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
+/** export upload plugin */
 export default {
     install(app: App) {
         registerComponent(app, Upload);
     },
 } as Plugin;
 
+/** export upload components */
 export { Upload as OUpload };
